e020 - New names, pyqt4 and pyqt5 support, and speed control
Changed names for the .ui files, their correspondent .py files and for
the .py files that implement the behaviour of the interfaces. Added a
check for the presence of PyQt5 (now the recommended version of PyQt).
Added also a couple of functions that translate the piezo speed to DAC
bits and back.