e033 - New options
Added a combo box to the hw config ui that allows to specify if the AFM
has the piezo actuator under the sample or above the tip. Added also a
couple of functions and parameters that convert the reduced range of the
monitor analog inputs (0-3.3V) to a range specified by the user in the
configuration file.